requirements.txt
tensorflow
pandas
numpy
scikit-learn
matplotlib
seaborn
fastapi
uvicorn
python-multipart
Pillow
